[Feat]: 媒体显示 MAL 评分功能

+++ lib-common/src/main/java/com/xiaoyv/common/api/parser/entity/MediaDetailEntity.kt
@@ -63,7 +63,7 @@ data class MediaDetailEntity(
     @Parcelize
     @Keep
     data class FriendRating(
-        var score: Double = 0.0,
+        var score: Float = 0f,
         var count: Int = 0,
         var desc: String = "",
     ) : Parcelable
